Various Cryo/Vivarium re-work Various UI updates.

+++ ARKViewer/Models/ContentDroppedItem.cs
@@ -22,7 +22,7 @@ namespace ARKViewer.Models
         [DataMember] public float? X { get; set; }
         [DataMember] public float? Y { get; set; }
         [DataMember] public float? Z { get; set; }
-        [DataMember] public int? InventoryId { get; set; }
+        [DataMember] public long? InventoryId { get; set; }
 
     }
 }
